Implement an ION allocator helper class
This helper class is used to implement the SDP Prototype in the OpenCDM
browser plugin. It allows to allocate the 'secure' ION buffer used to output
the decrypted content. The buffer is allocated in the specified secure ION
heap ID. Once the decryption is completed, it allows to map the ION buffer
and to copy the data to the shared memory. This assumes that the 'secure'
ION heap is CPU accessible meaning it is not actually protected by any
hardware mechanism such as TZASC.

The current implementation only supports the ION ABI as defined in kernel 4.9.

Signed-off-by: Alexandre Jutras <alexandre.jutras@linaro.org>