Implement a socket client helper class
This socket client helper class allows the OpenCDM browser plugin to send the
secure (ION) file decriptor to the CDMI Service. The socket server helper class
implements the service side of the communication.

The helper class provides an API to connect to a socket (typically when
creating a media engine session) and to send the file descriptor (typically
at every decrypt operation). The file descriptor is communicated through the
socket's ancillary data. The secure buffer size is communicated through the
socket's data.

Signed-off-by: Alexandre Jutras <alexandre.jutras@linaro.org>