Implement the OpenCDM SDP Prototype
The purpose of the OpenCDM SDP Prototype is to validate the integration of the
ION buffer into OpenCDM and into the DRM layer. This is used to demonstrate
the OpenCDM's part of the SDP implementation on platforms which cannot
implement the end-to-end SDP or on platforms which does not support it yet.

The prototype is implemented at the OpenCDM browser plugin level. At every
decryption operation, it allocate a 'secure' ION buffer from the ION heap
specified by the ION_SECURE_HEAP_ID_DECODER compile-time variable. Then, it
communicates the corresponding file descriptor to the DRM layer via the CDMI
service. The interprocess communication relies on a per-session socket
connection established between the OpenCDM browser plugin and the CDMI service.

Once the DRM implementation has completed the content decryption into the
'secure' buffer, the control is returned to the browser plugin. It maps the
memory referenced by ION buffer, copies the data to the shared memory and
passes that shared memory to the audio and video frameworks to proceed with
the audio/video pipeline. This mechanism assumes that the 'secure' ION heap is
CPU accessible meaning it is not actually protected by any hardware mechanism
such as TZASC.

This has been tested on NXP i.MX8M using PlayReady's OpenCDM integration in
Chromium.

Signed-off-by: Alexandre Jutras <alexandre.jutras@linaro.org>